requirements.txt
langchain
langchain-google-genai
langchain-community
google-generativeai
python-dotenv